Add nested extension for E3 timestamp experiment
Including this extension here is a kludge, meant to route around a delay
in Gerrit repo creation that is currently blocking the E3 team. The
LastModified extension itself has not been modified and its
functionality has not been impacted.

This experiment is shown to 50% of editors, on a sample factor of 0.006 of
articles. Those in the experiment will see the time the article was last
modified presented in a human-readable format in the article's header.
Clicking on the timestamp will log the event and redirect the user to
the article's history page.

Specific information regarding this experiment can be found on its Meta
page: https://meta.wikimedia.org/wiki/Research:Timestamp_position_modification

Change-Id: Ic57fc3cacbf68d21cb493260c8b566608c681c30