(bug 37255) Adjust position of timestamp as needed
The place where we drop in the timestamp may also contain various other
UI elements. We need to inspect the DOM and adjust the position of the
timestamp in light of what we discover.
As an extra precaution, make sure we aren't offsetting an element by
a value greater than its container's width.

Change-Id: I27bf0604b96e972c0a4ff7e8510751496a2f4793