Changing wgMainPageTitle to use wgPageName.
Before this change, the links to the history pages are always pointing
to the history page of "Main Page", which is rather annoying to users
that actually click on the link on any other page, but is being directed
to the wrong history page.

This change changes the link to use the variable for the current page
that a user is looking at and outputs the correct link for the user.

Change-Id: I8153358c6025108aec946db9b2cf8e904f838d1b